クリーチャーの追加をドラッグアンドドロップではなくボタンから行うように変更（タブレットでのタッチ操作に対応できなかったため）
