+++ components/new_counter_view.vue
@@ -45,7 +45,8 @@
         this.$emit('end');
       },
       getCounter(): Counter {
-        return this.counter;
+        //this.creatureをそのまま返すと、パワーとタフネスがstringになっている場合があるので作成しなおす
+        return Counter.create(this.counter.name, Number(this.counter.powerBonus), Number(this.counter.toughnessBonus))
       },
     }
   }
